update EV scraper for GA
some precincts don't report by mode and instead report a total, but it is duplicated 4x in the data. Fix now adjusts this number down and keeps the numeric estimate by mode in the files.

+++ scripts/functions.R
@@ -130,7 +130,7 @@ get_data <- function(state, county, timestamp, path = NULL) {
   )
  
   # Upstream fix to precinct_total
-  d$precinct_total <- as.integer(d$precinct_total)
+  d$precinct_total <- as.numeric(d$precinct_total)
 
   local_file_name <- ifelse(is.na(county), state, glue("{state}_{county}"))
   clarity_counties <- list(
